requirements.txt
tree-sitter-language-pack
python-dotenv
GitPython
networkx
sentence-transformers
torch
openai